UIU-1167: Revert replacing files
